requirements.txt
opendatasets
pandas
numpy
tensorflow
scikit-learn
Pillow
kagglehub
matplotlib
python-dotenv